TODO_project/requirements.txt
altgraph==0.17.2
aniso8601==7.0.0
asgiref==3.4.1
bookmarks2markdown==1.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.3.0
docker==5.0.3
drf-yasg==1.20.0
dynaconf==3.1.7
Faker==9.8.2
form==0.0.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lxml==4.6.3
macholib==1.15.2
MarkupSafe==2.0.1
mixer==7.2.0
oauthlib==3.1.1
packaging==21.3
path==16.2.0
Pillow==8.4.0
promise==2.3
psycopg2-binary==2.9.2
pycparser==2.21
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
PyJWT==2.3.0
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.3.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
websocket-client==1.2.3
